+++ WinQuake/screen.c
@@ -114,7 +114,7 @@ void SCR_EraseCenterString (void)
 	}
 
 	if (scr_center_lines <= 4)
-		y = vid.height*0.35;
+		y = vid.height*0.35f;
 	else
 		y = 48;
 
@@ -140,7 +140,7 @@ void SCR_DrawCenterString (void)
 	start = scr_centerstring;
 
 	if (scr_center_lines <= 4)
-		y = vid.height*0.35;
+		y = vid.height*0.35f;
 	else
 		y = 48;
 
@@ -200,9 +200,9 @@ float CalcFov (float fov_x, float width, float height)
         if (fov_x < 1 || fov_x > 179)
                 Sys_Error ("Bad fov: %f", fov_x);
 
-        x = width/tan(fov_x/360*M_PI);
+        x = width/tanf(fov_x/360*M_PI);
 
-        a = atan (height/x);
+        a = atanf (height/x);
 
         a = a*360/M_PI;
 
@@ -712,7 +712,7 @@ void SCR_DrawNotifyString (void)
 
 	start = scr_notifystring;
 
-	y = vid.height*0.35;
+	y = vid.height*0.35f;
 
 	do	
 	{
